pyproject.toml
[tool.poetry]
name = "fastapi-arango-backend"
version = "0.1.0"
description = "A reusable FastAPI backend with ArangoDB database integration."
authors = ["Jason Mangawang <jmangawang@gmail.com>"]
packages = [{ include = "backend" }]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.115.0"
python-dotenv = "^1.1.0"
uvicorn = "^0.34.0"
loguru = "^0.7.3"
jsonschema = "^4.23.0"
python-arango = "^8.1.6"
httpx = "^0.28.1"
pytest-ordering = "^0.6"
pytest-env = "0.8.2"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
flake8 = "^6.1.0"
pytest-asyncio = "0.21.1"

[build-system]
requires = ["poetry-core>=2.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.config]
virtualenvs.in-project = true
